Move out create room button to separate class
