+++ osu.Game/Screens/Multi/Multiplayer.cs
@@ -21,9 +21,7 @@ using osu.Game.Screens.Multi.Components;
 using osu.Game.Screens.Multi.Lounge;
 using osu.Game.Screens.Multi.Lounge.Components;
 using osu.Game.Screens.Multi.Match;
-using osu.Game.Screens.Multi.Match.Components;
 using osu.Game.Users;
-using osuTK;
 
 namespace osu.Game.Screens.Multi
 {
@@ -330,27 +328,6 @@ namespace osu.Game.Screens.Multi
             private class BackgroundSprite : UpdateableBeatmapBackgroundSprite
             {
                 protected override double TransformDuration => 200;
-            }
-        }
-
-        public class CreateRoomButton : PurpleTriangleButton
-        {
-            public CreateRoomButton()
-            {
-                Size = new Vector2(150, Header.HEIGHT - 20);
-                Margin = new MarginPadding
-                {
-                    Top = 10,
-                    Right = 10 + HORIZONTAL_OVERFLOW_PADDING,
-                };
-            }
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                Triangles.TriangleScale = 1.5f;
-
-                Text = "Create room";
             }
         }
     }
